wgpu: Move command renderer to surface/commands.rs
